requirements.txt
absl-py==1.0.0
astunparse==1.6.3
attrs==21.4.0
-e git+https://github.com/KillerX629/Gym-retro.git@8cce9ad5f8d7c4cf7cb8835bb667908e93abef81#egg=baselines&subdirectory=baselines/baselines-tf2
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
cloudpickle==1.2.2
dill==0.3.4
flatbuffers==2.0
future==0.18.2
gast==0.5.3
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
gym==0.13.1
gym-notices==0.0.6
gym-retro @ file:///home/ignacio/Documents/repos/Gym/gym_retro-0.8.0.dev3%2Bgf347d7e-cp38-cp38-linux_x86_64.whl
h5py==3.6.0
idna==3.3
importlib-metadata==4.11.3
iniconfig==1.1.1
joblib==1.1.0
keras==2.8.0
Keras-Preprocessing==1.1.2
libclang==13.0.0
Markdown==3.3.6
numpy==1.22.3
oauthlib==3.2.0
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
pluggy==1.0.0
progressbar2==4.0.0
protobuf==3.19.4
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyglet==1.3.2
pyparsing==3.0.7
python-utils==3.1.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.8.0
six==1.16.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-gpu==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
tomli==2.0.1
tqdm==4.63.0
typing_extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.0.3
wrapt==1.14.0
zipp==3.7.0

